# api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: spond@temple.edu
    url: http://github.com/veg/api-datamonkey
  description: |
    Datamonkey is a free public server for comparative analysis of sequence alignments using state-of-the-art statistical models. <br> This is the OpenAPI definition for the Datamonkey API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Datamonkey API
  version: 1.0.0
  x-logo:
    url: https://raw.githubusercontent.com/veg/datamonkey-js/refs/heads/master/public/assets/img/header-logo.svg
    altText: Datamonkey.org logo
...
servers:
- url: http://datamonkey.org/api/v1
- url: https://datamonkey.org/api/v1
...
paths:
...
  /methods/meme-start:
    post:
      operationId: startMEMEJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemeRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Success
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidDataError'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Start a MEME analysis
      tags:
      - MEME
  /methods/meme-result:
    post:
      operationId: getMEMEJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemeRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemeResult'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Get MEME analysis results
      tags:
      - MEME
  /methods/relax-start:
    post:
      operationId: startRELAXJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelaxRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Success
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidDataError'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Start and monitor a RELAX job
      tags:
      - RELAX
  /methods/relax-result:
    post:
      operationId: getRELAXJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelaxRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelaxResult'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Get RELAX analysis results
      tags:
      - RELAX
...
  /methods/multihit-start:
    post:
      description: "MULTI-HIT (Multiple Hit) examines whether or not a codon alignment\
        \ is better fit by models \nwhich permit multiple instantaneous substitutions.\
        \ It can analyze both double-hit and \ntriple-hit substitutions, with an option\
        \ to account for synonymous triple-hit substitutions.\n"
      operationId: startMULTIHITJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultihitRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Success
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidDataError'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Start a MULTI-HIT analysis
      tags:
      - MULTI-HIT
  /methods/multihit-result:
    post:
      description: |
        Retrieve the results of a previously submitted MULTI-HIT analysis job.
        Results include evidence ratios for different substitution types, site-specific substitutions,
        model fitting information, and statistical test results comparing different models.
      operationId: getMULTIHITJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultihitRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultihitResult'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Get MULTI-HIT analysis results
      tags:
      - MULTI-HIT
...
components:
  schemas:
...
    GeneticCode:
      default: Universal
      description: |
        Available Genetic Codes
      enum:
      - Universal
      - Vertebrate mtDNA
      - Yeast mtDNA
      - Mold/Protozoan mtDNA
      - Invertebrate mtDNA
      - Ciliate Nuclear
      - Echinoderm mtDNA
      - Eupltoid Nuclear
      - Alt. Yeast Nuclear
      - Ascidian mtDNA
      - Flatworm mtDNA
      - Blepharisma Nuclear
      type: string
...
    JobStatus:
      example:
        error_message: error_message
        job_id: job_id
        status: queued
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        status:
          enum:
          - queued
          - running
          - completed
          - error
          type: string
        error_message:
          type: string
      type: object
    InvalidDataError:
      example:
        errors:
        - field: field
          message: message
        - field: field
          message: message
      properties:
        errors:
          items:
            $ref: '#/components/schemas/InvalidDataError_errors_inner'
          type: array
      type: object
    ServerError:
      example:
        code: code
        details: details
        error: error
      properties:
        error:
          description: A general description of the error.
          type: string
        code:
          description: A unique error code (if available).
          type: string
        details:
          description: Optional details about the error.
          type: string
      type: object
...
    MemeRequest:
      example:
        genetic_code: Universal
        impute_states: "No"
        site_multihit: Estimate
        multiple_hits: None
        rates: 0
        tree: tree
        alignment: alignment
        resample: 6.027456183070403
      properties:
        alignment:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        tree:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        genetic_code:
          $ref: '#/components/schemas/GeneticCode'
        multiple_hits:
          default: None
          description: Specify handling of multiple nucleotide substitutions
          enum:
          - None
          - Double
          - Double+Triple
          type: string
        site_multihit:
          default: Estimate
          description: Specify whether to estimate multiple hit rates for each site
          enum:
          - Estimate
          - Global
          type: string
        rates:
          default: 2
          description: Number of different categories of non-synonymous rates
          type: integer
        resample:
          default: 0
          description: Number of bootstrapping replicates
          type: number
        impute_states:
          default: "No"
          description: Option to impute likely character states for missing data
          enum:
          - "Yes"
          - "No"
          type: string
      type: object
    MemeResult:
      example:
        result:
          input:
            trees:
              key: trees
          timers:
            key:
              timer: 6
              order: 0
          data partitions:
            key:
              coverage:
              - - 1
                - 1
              - - 1
                - 1
              name: name
          branch attributes:
            attributes:
              key:
                display order: 4
                attribute type: attribute type
            additionalProperties:
              key:
                Global MG94xREV: 3.616076749251911
                Nucleotide GTR: 2.027123023002322
                original name: original name
          tested:
            key:
              key: tested
          MLE:
            headers:
            - - headers
              - headers
            - - headers
              - headers
            content:
              key:
              - - 7.386281948385884
                - 7.386281948385884
              - - 7.386281948385884
                - 7.386281948385884
          runtime: runtime
          analysis:
            settings:
              site_filter: "{}"
            contact: john.smith@example.com
          fits:
            key:
              Log Likelihood: 2.3021358869347655
              estimated parameters: 9.301444243932576
              Rate Distributions: "{}"
              AIC-c: 5.962133916683182
              display order: 7.061401241503109
              Equilibrium frequencies:
              - - 5.637376656633329
                - 5.637376656633329
              - - 5.637376656633329
                - 5.637376656633329
        job_id: job_id
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        result:
          $ref: '#/components/schemas/FelResult_result'
      type: object
    RelaxRequest:
      example:
        genetic_code: Universal
        models: All
        reference_branches:
        - reference_branches
        - reference_branches
        kill_zero_lengths: "No"
        test_branches:
        - test_branches
        - test_branches
        rates: 0
        tree: tree
        alignment: alignment
      properties:
        alignment:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        tree:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        genetic_code:
          $ref: '#/components/schemas/GeneticCode'
        test_branches:
          default: []
          description: Branches to be considered as 'Test'
          items:
            type: string
          type: array
        reference_branches:
          default: []
          description: Branches to be considered as 'Reference'
          items:
            type: string
          type: array
        models:
          default: All
          description: "Type of analysis to run (All for descriptive models and RELAX\
            \ test, Minimal for only the RELAX test)"
          enum:
          - All
          - Minimal
          type: string
        rates:
          default: 3
          description: Number of omega rate classes
          type: integer
        kill_zero_lengths:
          default: "No"
          description: Specify whether to handle zero-length branches
          enum:
          - "Yes"
          - "No"
          type: string
      type: object
    RelaxResult:
      example:
        result:
          input:
            trees:
              key: trees
          timers:
            key:
              timer: 6
              order: 0
          data partitions:
            key:
              coverage:
              - - 1
                - 1
              - - 1
                - 1
              name: name
          test results:
            branch attributes:
              key:
                key:
                  RELAX null: 6.778324963048013
                  k (RELAX alternative): 5.944895607614016
                  Nucleotide GTR: 3.5571952270680973
                  tested: true
                  MG94xREV with separate rates for branch sets: 6.965117697638846
                  original name: original name
                  General descriptive: 1.284659006116532
                  RELAX alternative: 6.878052220127876
                  k (general descriptive): 2.8841621266687802
            p-value: 6.683562403749608
            test: test
            "null": "null"
            log likelihood ratio: 9.018348186070783
            relaxation or intensification parameter: 8.762042012749001
            background: background
            degrees of freedom: 6
            alternative: alternative
          tested:
            key:
              key: tested
          runtime: runtime
          analysis:
            settings:
              site_filter: "{}"
            contact: john.smith@example.com
          fits:
            RELAX null:
              Log Likelihood: 4.965218492984954
              estimated parameters: 5
              Rate Distributions:
                Test:
                  key:
                    proportion: 7.457744773683766
                    omega: 6.84685269835264
                Reference:
                  key:
                    proportion: 7.457744773683766
                    omega: 6.84685269835264
              AIC-c: 9.965781217890562
              display order: 9
            Nucleotide GTR:
              Log Likelihood: 0.8008281904610115
              estimated parameters: 6
              Rate Distributions:
                key: 5.637376656633329
              AIC-c: 1.4658129805029452
              display order: 2
              Equilibrium frequencies:
              - - 5.962133916683182
                - 5.962133916683182
              - - 5.962133916683182
                - 5.962133916683182
            MG94xREV with separate rates for branch sets:
              Log Likelihood: 7.061401241503109
              estimated parameters: 9
              Rate Distributions:
                key:
                - - 4.145608029883936
                  - 4.145608029883936
                - - 4.145608029883936
                  - 4.145608029883936
              AIC-c: 3.616076749251911
              display order: 7
              Equilibrium frequencies:
              - - 2.027123023002322
                - 2.027123023002322
              - - 2.027123023002322
                - 2.027123023002322
            General descriptive:
              Log Likelihood: 1.2315135367772556
              estimated parameters: 1
              Rate Distributions:
                Shared:
                  key:
                    proportion: 7.457744773683766
                    omega: 6.84685269835264
              AIC-c: 1.4894159098541704
              display order: 1
            RELAX alternative:
              Log Likelihood: 4.965218492984954
              estimated parameters: 5
              Rate Distributions:
                Test:
                  key:
                    proportion: 7.457744773683766
                    omega: 6.84685269835264
                Reference:
                  key:
                    proportion: 7.457744773683766
                    omega: 6.84685269835264
              AIC-c: 9.965781217890562
              display order: 9
        job_id: job_id
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        result:
          $ref: '#/components/schemas/RelaxResult_result'
      type: object
...
    MultihitRequest:
      description: Request parameters for MULTI-HIT analysis
      example:
        genetic_code: Universal
        triple_islands: false
        alignment_file: alignment_file
        rate_classes: 1
      properties:
        alignment_file:
          description: "Path to the alignment file (e.g., .fasta, .phy)"
          type: string
        genetic_code:
          default: Universal
          description: The genetic code to use for the analysis
          type: string
        triple_islands:
          default: false
          description: Toggle for accounting synonymous triple-hit substitutions
          type: boolean
        rate_classes:
          default: 3
          description: Number of rate classes to use
          maximum: 10
          minimum: 1
          type: integer
      required:
      - alignment_file
      - genetic_code
      type: object
    MultihitResult:
      description: Results from a MULTI-HIT analysis
      example:
        result:
          site_substitutions:
            key:
              key: "{}"
          evidence_ratios:
            three_hit_islands_vs_2_hit:
            - - 1.4658129805029452
              - 1.4658129805029452
            - - 1.4658129805029452
              - 1.4658129805029452
            three_hit:
            - - 0.8008281904610115
              - 0.8008281904610115
            - - 0.8008281904610115
              - 0.8008281904610115
            two_hit:
            - - 6.027456183070403
              - 6.027456183070403
            - - 6.027456183070403
              - 6.027456183070403
          analysis:
            requirements: requirements
            contact: contact
            version: version
            authors: authors
            info: info
          fits:
            Standard_MG94:
              display_order: 2
              AIC_c: 5.962133916683182
              estimated_parameters: 7
              Log_Likelihood: 5.637376656633329
              Rate_Distributions: "{}"
            MG94_with_double_and_triple_instantaneous_substitutions_only_synonymous_islands:
              display_order: 2
              AIC_c: 5.962133916683182
              estimated_parameters: 7
              Log_Likelihood: 5.637376656633329
              Rate_Distributions: "{}"
            MG94_with_double_and_triple_instantaneous_substitutions:
              display_order: 2
              AIC_c: 5.962133916683182
              estimated_parameters: 7
              Log_Likelihood: 5.637376656633329
              Rate_Distributions: "{}"
            MG94_with_double_instantaneous_substitutions:
              display_order: 2
              AIC_c: 5.962133916683182
              estimated_parameters: 7
              Log_Likelihood: 5.637376656633329
              Rate_Distributions: "{}"
            Nucleotide_GTR:
              display_order: 2
              AIC_c: 5.962133916683182
              estimated_parameters: 7
              Log_Likelihood: 5.637376656633329
              Rate_Distributions: "{}"
          test_results:
            Double_hit_vs_single_hit:
              p_value: 3.616076749251911
              LRT: 9.301444243932576
            Triple_hit_vs_Triple_hit_island:
              p_value: 3.616076749251911
              LRT: 9.301444243932576
            Triple_hit_vs_double_hit:
              p_value: 3.616076749251911
              LRT: 9.301444243932576
            Triple_hit_vs_single_hit:
              p_value: 3.616076749251911
              LRT: 9.301444243932576
            Triple_hit_island_vs_double_hit:
              p_value: 3.616076749251911
              LRT: 9.301444243932576
        job_id: job_id
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        result:
          $ref: '#/components/schemas/MultihitResult_result'
      type: object
...
    InvalidDataError_errors_inner:
      example:
        field: field
        message: message
      properties:
        field:
          description: The field that caused the error
          type: string
        message:
          description: The error message
          type: string
      type: object
...
    FelResult_result:
      allOf:
      - $ref: '#/components/schemas/BaseResponse'
      - properties:
          fits:
            additionalProperties:
              $ref: '#/components/schemas/FelResult_result_allOf_fits_value'
            type: object
          branch attributes:
            $ref: '#/components/schemas/FelResult_result_allOf_branch_attributes'
          MLE:
            $ref: '#/components/schemas/FelResult_result_allOf_MLE'
        type: object
      example:
        input:
          trees:
            key: trees
        timers:
          key:
            timer: 6
            order: 0
        data partitions:
          key:
            coverage:
            - - 1
              - 1
            - - 1
              - 1
            name: name
        branch attributes:
          attributes:
            key:
              display order: 4
              attribute type: attribute type
          additionalProperties:
            key:
              Global MG94xREV: 3.616076749251911
              Nucleotide GTR: 2.027123023002322
              original name: original name
        tested:
          key:
            key: tested
        MLE:
          headers:
          - - headers
            - headers
          - - headers
            - headers
          content:
            key:
            - - 7.386281948385884
              - 7.386281948385884
            - - 7.386281948385884
              - 7.386281948385884
        runtime: runtime
        analysis:
          settings:
            site_filter: "{}"
          contact: john.smith@example.com
        fits:
          key:
            Log Likelihood: 2.3021358869347655
            estimated parameters: 9.301444243932576
            Rate Distributions: "{}"
            AIC-c: 5.962133916683182
            display order: 7.061401241503109
            Equilibrium frequencies:
            - - 5.637376656633329
              - 5.637376656633329
            - - 5.637376656633329
              - 5.637376656633329
...